+++ .eslintrc.js
@@ -1,6 +1,3 @@
 module.exports = {
-  extends: [
-    './packages/eslint-config-encode/index.js',
-    './packages/eslint-config-encode/vue.js',
-  ],
+  extends: ['./packages/eslint-config-encode/index.js', './packages/eslint-config-encode/vue.js'],
 }