+++ packages/encode-lint-cli/src/cli.ts
@@ -73,7 +73,7 @@ program
     const checking = ora()
     checking.start(`执行 ${PKG_NAME} 代码检查`)
 
-    console.log('11111111111111111111')
+    console.log('222')
 
     const { results, errorCount, warningCount, runErrors } = await scan({
       cwd,
@@ -91,7 +91,7 @@ program
       type = 'warn'
     }
 
-    console.log('🚀 ~ .action ~ runErrors:', runErrors)
+    // console.log('🚀 ~ .action ~ runErrors:', runErrors)
 
     checking[type]()
     if (results.length > 0) printReport(results, false)