+++ packages/encode-lint-cli/CHANGELOG.md
@@ -1,5 +1,16 @@
 # encode-lint-cli
 
+## 2.1.0
+
+### Minor Changes
+
+- 更新ESLint配置和支持包至null-eslint-config-encode
+
+### Patch Changes
+
+- Updated dependencies
+  - null-eslint-config-encode@2.1.0
+
 ## 2.0.0
 
 ### Major Changes