+++ packages/eslint-config-encode/CHANGELOG.md
@@ -1,5 +1,11 @@
 # null-eslint-config-encode
 
+## 2.1.0
+
+### Minor Changes
+
+- 更新ESLint配置和支持包至null-eslint-config-encode
+
 ## 2.0.0
 
 ### Major Changes